Fix sync of ports for 0 state
Original code only sync'd if the delta between last sent and current value was more than 0.001.
If a port was zeroed e.g. water draining from a cylinder, but the change was less than 0.001, then the client would not receive an update.

New method uses the same delta parameters but also checks if the new state is zero and the last sent state was non-zero.