Prevent client couplers breaking in a derail
During derail the base game changes the way the joints behave, resulting in inconsistent simulation between the server and client(s).

`TrainCar.UpdateCouplerJoints` is called on Derail, by patching it so it's only active on the server, the clients are far less likely to break couplings. Broken couplings on the server are communicated with the TrainCarUncouple packet.